Assignment 7

See merge request kevinkarnani/kk3286-kevin-karnani-se-181-course-project!2

+++ src/main/java/banking/Bank.java
@@ -17,7 +17,7 @@ public class Bank {
         this.accounts.put(uuid, new SavingsAccount(apr));
     }
 
-    public void createCDAccount(int uuid, int amount, double apr) {
+    public void createCDAccount(int uuid, double amount, double apr) {
         this.accounts.put(uuid, new CDAccount(amount, apr));
     }
