+++ src/main/java/banking/CDAccount.java
@@ -1,7 +1,7 @@
 package banking;
 
 public class CDAccount extends Account {
-    public CDAccount(int amount, double apr) {
+    public CDAccount(double amount, double apr) {
         this.amount = amount;
         this.apr = apr;
         this.maxDeposit = -1;
